+++ api/v2/podFinder_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"testing"
 
+	cloudoguerrors "github.com/cloudogu/ces-commons-lib/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -55,6 +56,20 @@ func TestGetPodForLabels(t *testing.T) {
 		// then
 		require.Error(t, err)
 		assert.ErrorContains(t, err, "failed to get pods")
+	})
+	t.Run("should return an error when no pod was found", func(t *testing.T) {
+		// given
+		labels := CesMatchingLabels{DoguLabelName: "ldap", DoguLabelVersion: "1.2.3-4"}
+		cli := newMockK8sClient(t)
+		cli.On("List", testCtx, mock.Anything, client.MatchingLabels(labels)).Return(nil)
+
+		// when
+		_, err := GetPodForLabels(testCtx, cli, labels)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "found no pods for labels")
+		assert.True(t, cloudoguerrors.IsNotFoundError(err))
 	})
 	t.Run("should return for multiple pods for unique labels", func(t *testing.T) {
 		// given