+++ controllers/resource/resource_generator_test.go
@@ -248,13 +248,14 @@ func TestResourceGenerator_GetDoguService(t *testing.T) {
 	t.Run("Return simple service", func(t *testing.T) {
 		// given
 		ldapDoguResource := readLdapDoguResource(t)
+		ldapDogu := readLdapDogu(t)
 		imageConf := readLdapDoguImageConfig(t)
 		generator := resourceGenerator{
 			scheme: getTestScheme(),
 		}
 
 		// when
-		actualService, err := generator.CreateDoguService(ldapDoguResource, imageConf)
+		actualService, err := generator.CreateDoguService(ldapDoguResource, ldapDogu, imageConf)
 
 		assert.NoError(t, err)
 		assert.Equal(t, readLdapDoguExpectedService(t), actualService)
@@ -267,6 +268,7 @@ func TestResourceGenerator_GetDoguService(t *testing.T) {
 		}
 
 		ldapDoguResource := readLdapDoguResource(t)
+		ldapDogu := readLdapDogu(t)
 		imageConf := readLdapDoguImageConfig(t)
 
 		oldMethod := ctrl.SetControllerReference
@@ -276,7 +278,7 @@ func TestResourceGenerator_GetDoguService(t *testing.T) {
 		defer func() { ctrl.SetControllerReference = oldMethod }()
 
 		// when
-		_, err := generator.CreateDoguService(ldapDoguResource, imageConf)
+		_, err := generator.CreateDoguService(ldapDoguResource, ldapDogu, imageConf)
 
 		// then
 		require.Error(t, err)